fix: split mock_factory session helpers to satisfy backend-size limit
